+++ puppyfw.php
@@ -7,7 +7,7 @@
  * Author URI:      https://truongwp.com/
  * Text Domain:     puppyfw
  * Domain Path:     /languages
- * Version:         0.2.0
+ * Version:         0.3.0
  *
  * @package         PuppyFW
  */