+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## [Unreleased]
+## [0.3.0]
 ### Added
 - Options page builder
 - Hook `puppyfw_before_init`
@@ -40,7 +41,8 @@
 ## [0.1.0] - 2017-09-19
 - First release
 
-[Unreleased]: https://github.com/truongwp/puppyfw/compare/0.2.0...HEAD
+[Unreleased]: https://github.com/truongwp/puppyfw/compare/0.3.0...HEAD
+[0.3.0]: https://github.com/truongwp/puppyfw/compare/0.2.0...HEAD
 [0.2.0]: https://github.com/truongwp/puppyfw/compare/0.1.2...HEAD
 [0.1.2]: https://github.com/truongwp/puppyfw/compare/0.1.1...0.1.2
 [0.1.1]: https://github.com/truongwp/puppyfw/compare/0.1.0...0.1.1